# roadrecon/frontend-react/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "ROADrecon",
  "version": "2.0.0"
 },
 "paths": {
...
  "/api/azure-role-assignments": {
   "get": {
    "tags": [
     "governance"
    ],
    "summary": "List Azure Role Assignments",
    "operationId": "list_azure_role_assignments_api_azure_role_assignments_get",
    "parameters": [
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "page_size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 500,
       "minimum": 1,
       "default": 50,
       "title": "Page Size"
      }
     },
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "name": "sort",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Sort"
      }
     },
     {
      "name": "order",
      "in": "query",
      "required": false,
      "schema": {
       "enum": [
        "asc",
        "desc"
       ],
       "type": "string",
       "default": "asc",
       "title": "Order"
      }
     },
     {
      "name": "filter",
      "in": "query",
      "required": false,
      "schema": {
       "type": "array",
       "items": {
        "type": "string"
       },
       "description": "Advanced filters, repeatable: `field:op:value`. For `in`/`notIn` the value is a comma list of URI-encoded items.",
       "default": [],
       "title": "Filter"
      },
      "description": "Advanced filters, repeatable: `field:op:value`. For `in`/`notIn` the value is a comma list of URI-encoded items."
     },
     {
      "name": "match",
      "in": "query",
      "required": false,
      "schema": {
       "enum": [
        "all",
        "any"
       ],
       "type": "string",
       "description": "How filters combine.",
       "default": "all",
       "title": "Match"
      },
      "description": "How filters combine."
     },
     {
      "name": "principalId",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Principalid"
      }
     },
     {
      "name": "transitive",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "boolean"
        },
        {
         "type": "null"
        }
       ],
       "title": "Transitive"
      }
     },
     {
      "name": "kind",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "enum": [
          "active",
          "eligible"
         ],
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Kind"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Page_AzureRoleAssignmentRow_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/pim-assignments": {
   "get": {
    "tags": [
     "governance"
    ],
    "summary": "List Pim Assignments",
    "operationId": "list_pim_assignments_api_pim_assignments_get",
    "parameters": [
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "page_size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 500,
       "minimum": 1,
       "default": 50,
       "title": "Page Size"
      }
     },
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "name": "sort",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Sort"
      }
     },
     {
      "name": "order",
      "in": "query",
      "required": false,
      "schema": {
       "enum": [
        "asc",
        "desc"
       ],
       "type": "string",
       "default": "asc",
       "title": "Order"
      }
     },
     {
      "name": "filter",
      "in": "query",
      "required": false,
      "schema": {
       "type": "array",
       "items": {
        "type": "string"
       },
       "description": "Advanced filters, repeatable: `field:op:value`. For `in`/`notIn` the value is a comma list of URI-encoded items.",
       "default": [],
       "title": "Filter"
      },
      "description": "Advanced filters, repeatable: `field:op:value`. For `in`/`notIn` the value is a comma list of URI-encoded items."
     },
     {
      "name": "match",
      "in": "query",
      "required": false,
      "schema": {
       "enum": [
        "all",
        "any"
       ],
       "type": "string",
       "description": "How filters combine.",
       "default": "all",
       "title": "Match"
      },
      "description": "How filters combine."
     },
     {
      "name": "principalId",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Principalid"
      }
     },
     {
      "name": "transitive",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "boolean"
        },
        {
         "type": "null"
        }
       ],
       "title": "Transitive"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Page_PimAssignmentRow_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/groups/{id}/pim": {
   "get": {
    "tags": [
     "governance"
    ],
    "summary": "Get Group Pim",
    "description": "PIM for Groups settings and assignments; null when the group is not onboarded.",
    "operationId": "get_group_pim_api_groups__id__pim_get",
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "anyOf": [
          {
           "$ref": "#/components/schemas/GroupPim"
          },
          {
           "type": "null"
          }
         ],
         "title": "Response Get Group Pim Api Groups  Id  Pim Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/access-package-policies": {
   "get": {
    "tags": [
     "governance"
    ],
    "summary": "List Access Package Policies",
    "description": "Access package policies the user can request, directly or through a group.",
    "operationId": "list_access_package_policies_api_access_package_policies_get",
    "parameters": [
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "page_size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 500,
       "minimum": 1,
       "default": 50,
       "title": "Page Size"
      }
     },
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "name": "sort",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Sort"
      }
     },
     {
      "name": "order",
      "in": "query",
      "required": false,
      "schema": {
       "enum": [
        "asc",
        "desc"
       ],
       "type": "string",
       "default": "asc",
       "title": "Order"
      }
     },
     {
      "name": "filter",
      "in": "query",
      "required": false,
      "schema": {
       "type": "array",
       "items": {
        "type": "string"
       },
       "description": "Advanced filters, repeatable: `field:op:value`. For `in`/`notIn` the value is a comma list of URI-encoded items.",
       "default": [],
       "title": "Filter"
      },
      "description": "Advanced filters, repeatable: `field:op:value`. For `in`/`notIn` the value is a comma list of URI-encoded items."
     },
     {
      "name": "match",
      "in": "query",
      "required": false,
      "schema": {
       "enum": [
        "all",
        "any"
       ],
       "type": "string",
       "description": "How filters combine.",
       "default": "all",
       "title": "Match"
      },
      "description": "How filters combine."
     },
     {
      "name": "userId",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Userid"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Page_AccessPackagePolicyRow_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
   "AccessPackagePolicyRow": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "packageName": {
      "type": "string",
      "title": "Packagename"
     },
     "packageDescription": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Packagedescription"
     },
     "policyName": {
      "type": "string",
      "title": "Policyname"
     },
     "resources": {
      "items": {
       "$ref": "#/components/schemas/AccessPackageResource"
      },
      "type": "array",
      "title": "Resources"
     },
     "approvalRequired": {
      "type": "boolean",
      "title": "Approvalrequired"
     },
     "approvers": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Approvers"
     },
     "durationDays": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Durationdays"
     },
     "renewable": {
      "type": "boolean",
      "title": "Renewable"
     },
     "via": {
      "$ref": "#/components/schemas/ObjectRef",
      "description": "Why the user can request it: direct, a group, or a keyword like \"All members\"."
     }
    },
    "type": "object",
    "required": [
     "id",
     "packageName",
     "packageDescription",
     "policyName",
     "resources",
     "approvalRequired",
     "approvers",
     "durationDays",
     "renewable",
     "via"
    ],
    "title": "AccessPackagePolicyRow"
   },
   "AccessPackageResource": {
    "properties": {
     "kind": {
      "type": "string",
      "title": "Kind"
     },
     "resource": {
      "$ref": "#/components/schemas/ObjectRef"
     },
     "role": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Role"
     }
    },
    "type": "object",
    "required": [
     "kind",
     "resource",
     "role"
    ],
    "title": "AccessPackageResource"
   },
...
   "AzureRole": {
    "properties": {
     "displayName": {
      "type": "string",
      "title": "Displayname"
     },
     "description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     },
     "isBuiltIn": {
      "type": "boolean",
      "title": "Isbuiltin"
     }
    },
    "type": "object",
    "required": [
     "displayName",
     "description",
     "isBuiltIn"
    ],
    "title": "AzureRole"
   },
   "AzureRoleAssignmentRow": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "kind": {
      "type": "string",
      "enum": [
       "active",
       "eligible"
      ],
      "title": "Kind"
     },
     "role": {
      "$ref": "#/components/schemas/AzureRole"
     },
     "principal": {
      "$ref": "#/components/schemas/ObjectRef"
     },
     "via": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/ObjectRef"
       },
       {
        "type": "null"
       }
      ]
     },
     "scope": {
      "type": "string",
      "title": "Scope"
     },
     "conditional": {
      "type": "boolean",
      "title": "Conditional"
     }
    },
    "type": "object",
    "required": [
     "id",
     "kind",
     "role",
     "principal",
     "via",
     "scope",
     "conditional"
    ],
    "title": "AzureRoleAssignmentRow"
   },
...
   "GroupPim": {
    "properties": {
     "onboardedDateTime": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Onboardeddatetime"
     },
     "memberApprovalRequired": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Memberapprovalrequired"
     },
     "ownerApprovalRequired": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ownerapprovalrequired"
     },
     "members": {
      "items": {
       "$ref": "#/components/schemas/PimSubject"
      },
      "type": "array",
      "title": "Members"
     },
     "owners": {
      "items": {
       "$ref": "#/components/schemas/PimSubject"
      },
      "type": "array",
      "title": "Owners"
     }
    },
    "type": "object",
    "required": [
     "onboardedDateTime",
     "memberApprovalRequired",
     "ownerApprovalRequired",
     "members",
     "owners"
    ],
    "title": "GroupPim"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ObjectRef": {
    "properties": {
     "id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Id"
     },
     "type": {
      "type": "string",
      "enum": [
       "user",
       "group",
       "device",
       "servicePrincipal",
       "application",
       "administrativeUnit",
       "role",
       "policy",
       "namedLocation",
       "keyword",
       "value",
       "unknown"
      ],
      "title": "Type"
     },
     "displayName": {
      "type": "string",
      "title": "Displayname"
     },
     "sub": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Sub",
      "description": "Secondary identifier: UPN for users, appId for apps and SPs."
     }
    },
    "type": "object",
    "required": [
     "id",
     "type",
     "displayName"
    ],
    "title": "ObjectRef",
    "description": "Every reference to an object, anywhere. `keyword`/`value`/`unknown` are not links."
   },
   "Page_AccessPackagePolicyRow_": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/AccessPackagePolicyRow"
      },
      "type": "array",
      "title": "Items"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "page": {
      "type": "integer",
      "title": "Page"
     },
     "page_size": {
      "type": "integer",
      "title": "Page Size"
     }
    },
    "type": "object",
    "required": [
     "items",
     "total",
     "page",
     "page_size"
    ],
    "title": "Page[AccessPackagePolicyRow]"
   },
...
   "Page_AzureRoleAssignmentRow_": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/AzureRoleAssignmentRow"
      },
      "type": "array",
      "title": "Items"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "page": {
      "type": "integer",
      "title": "Page"
     },
     "page_size": {
      "type": "integer",
      "title": "Page Size"
     }
    },
    "type": "object",
    "required": [
     "items",
     "total",
     "page",
     "page_size"
    ],
    "title": "Page[AzureRoleAssignmentRow]"
   },
...
   "Page_PimAssignmentRow_": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/PimAssignmentRow"
      },
      "type": "array",
      "title": "Items"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "page": {
      "type": "integer",
      "title": "Page"
     },
     "page_size": {
      "type": "integer",
      "title": "Page Size"
     }
    },
    "type": "object",
    "required": [
     "items",
     "total",
     "page",
     "page_size"
    ],
    "title": "Page[PimAssignmentRow]"
   },
...
   "PimAssignmentRow": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "kind": {
      "type": "string",
      "enum": [
       "active",
       "eligible"
      ],
      "title": "Kind"
     },
     "resourceType": {
      "type": "string",
      "enum": [
       "directoryRole",
       "group",
       "other"
      ],
      "title": "Resourcetype"
     },
     "resource": {
      "$ref": "#/components/schemas/ObjectRef"
     },
     "role": {
      "type": "string",
      "title": "Role",
      "description": "Role name, or Member / Owner for groups."
     },
     "via": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/ObjectRef"
       },
       {
        "type": "null"
       }
      ]
     },
     "approvalRequired": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Approvalrequired"
     },
     "startDateTime": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Startdatetime"
     },
     "endDateTime": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Enddatetime",
      "description": "null = permanent."
     }
    },
    "type": "object",
    "required": [
     "id",
     "kind",
     "resourceType",
     "resource",
     "role",
     "via",
     "approvalRequired",
     "startDateTime",
     "endDateTime"
    ],
    "title": "PimAssignmentRow"
   },
   "PimSubject": {
    "properties": {
     "subject": {
      "$ref": "#/components/schemas/ObjectRef"
     },
     "kind": {
      "type": "string",
      "enum": [
       "active",
       "eligible"
      ],
      "title": "Kind"
     },
     "startDateTime": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Startdatetime"
     },
     "endDateTime": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Enddatetime"
     }
    },
    "type": "object",
    "required": [
     "subject",
     "kind",
     "startDateTime",
     "endDateTime"
    ],
    "title": "PimSubject"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}